+++ app/(tabs)/projects/_layout.tsx
@@ -1,5 +1,24 @@
 import { Colors } from "@/constants/Colors";
-import { Stack } from "expo-router";
+import { HeaderButton } from "@react-navigation/elements";
+import { router, Stack } from "expo-router";
+import React from "react";
+import { QrCodeIcon } from "react-native-heroicons/solid";
+
+const ProjectQrScan = () => (
+  <HeaderButton
+    accessibilityLabel="QR Scan"
+    onPress={() =>
+      router.push({
+        pathname: "/qrScan",
+        params: {
+          context: "project",
+        },
+      })
+    }
+  >
+    <QrCodeIcon size={22} color={"#fff"} />
+  </HeaderButton>
+);
 
 export default function TabsStackLayout() {
   return (
@@ -10,6 +29,7 @@ export default function TabsStackLayout() {
         headerTitleStyle: { color: "white" },
         headerTintColor: "white",
         headerTitle: "All Projects",
+        headerRight: () => <ProjectQrScan />
       }}
     />
   );
